pyproject.toml
[tool.poetry]
name = "python-api-challenge"
version = "0.1.0"
description = "Backend logic implementation with awesome FastAPI"
authors = [""]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "^0.6.1"
gunicorn = "^19.9"
fastapi = "^0.13.0"
pydantic = {version = "^0.21.0",extras = ["email"]}
passlib = {version = "^1.7",extras = ["bcrypt"]}
pyjwt = "^1.7"
asyncpg = "^0.18.3"
alembic = "^1.0"
psycopg2-binary = "^2.8"
python-slugify = "^3.0"
Unidecode = "^1.0"
python-dotenv = "^0.10.1"
databases = "^0.2.1"


[tool.poetry.dev-dependencies]

[tool.black]
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | alembic
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

